# backend/swagger/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Tarot API Test",
    "description": "오늘도 화이팅",
    "version": "1.0.0"
  },
  "host": "43.202.208.226:3001",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "JWT Access Token 입력 : 'Bearer {token}'"
    }
  },
  "paths": {
    "/tarot/guide": {
      "get": {
        "tags": [
          "Tarot"
        ],
        "summary": "가이드라인 불러오기",
        "description": "가이드라인 불러오기",
        "parameters": [
          {
            "name": "luckType",
            "in": "query",
            "description": "운 종류",
            "required": true,
            "type": "string",
            "example": "애정운",
            "value": "test_luck"
          },
          {
            "name": "luckOpt",
            "in": "query",
            "description": "멀티 여부",
            "required": true,
            "type": "integer",
            "example": "0",
            "value": "0"
          }
        ],
        "responses": {
          "200": {
            "description": "테스트 값 조회 성공"
          },
          "400": {
            "description": "잘못된 요청"
          }
        }
      }
    },
    "/tarot/poll/create": {
      "get": {
        "tags": [
          "Tarot"
        ],
        "summary": "타로 시작 시 Poll(임시저장) → 타로 시작 할 경우 뽑은 카드와 결과 저장을 구별할 Poll Table",
        "description": "뽑은 카드 결과 저장 및 총 결과 저장 시 사용됨",
        "parameters": [
          {
            "name": "userid",
            "in": "query",
            "description": "사용자의 ID",
            "required": true,
            "type": "string",
            "example": "Test ID = 1",
            "value": "1"
          }
        ],
        "responses": {
          "200": {
            "description": "Poll ID가 성공적으로 생성되었을 때의 응답",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "statusCode": {
                  "type": "number",
                  "example": 200
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Poll ID 생성 완료"
                    },
                    "pollId": {
                      "type": "number",
                      "example": 0
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "요청된 사용자 ID가 누락되었을 때의 응답",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "유저 아이디 누락"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "DB 저장 과정에서 오류 발생 시의 응답",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "DB 저장 오류"
                },
                "error": {
                  "type": "string",
                  "example": "에러 내용"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/tarot/card/info": {
      "post": {
        "tags": [
          "Tarot"
        ],
        "summary": "카드 정보 조회",
        "description": "카드 정보를 정수형으로 전달하면 해당 카드의 정보를 반환함",
        "parameters": [
          {
            "name": "card",
            "in": "query",
            "description": "카드 번호",
            "required": true,
            "type": "integer",
            "example": "1",
            "schema": {
              "type": "object",
              "properties": {
                "card": {
                  "type": "number",
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "카드 정보 불러오기 성공 url 전송",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "creat image url successfully"
                },
                "name": {
                  "type": "string",
                  "example": "카드 이름"
                },
                "english": {
                  "type": "string",
                  "example": "카드 영어 이름"
                },
                "mean": {
                  "type": "string",
                  "example": "카드 뜻"
                },
                "image_url": {
                  "type": "string",
                  "example": "카드 이미지 url"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "잘못된 카드 번호 요청",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "카드 번호가 없습니다."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "카드 정보 불러오기 실패",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "카드 정보를 가져오는데 실패했습니다."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
    "/user/signup": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "회원가입",
        "description": "회원가입 하는 유저의 아이디, username 중복검사 및 가입 하는 유저의 정보를 데이터베이스에 저장",
        "parameters": [
          {
            "name": "login_id",
            "in": "query",
            "description": "사용자의 아이디",
            "required": true,
            "type": "string",
            "example": "minki"
          },
          {
            "name": "password",
            "in": "query",
            "description": "비밀번호",
            "required": true,
            "type": "string",
            "example": "0000"
          },
          {
            "name": "name",
            "in": "query",
            "description": "닉네임",
            "required": true,
            "type": "string",
            "example": "타로마스터"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "접근 방식 오류"
          },
          "409": {
            "description": "이미 존재하는 아이디나 username"
          },
          "500": {
            "description": "회원가입 정보 불러오기 실패"
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "로그인",
        "description": "아이디와 비밀번호를 이용하여 로그인을 수행하고, 성공 시 액세스 토큰과 리프레시 토큰을 발급",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "사용자의 아이디",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "login_id": {
                  "type": "string",
                  "example": " "
                },
                "password": {
                  "type": "string",
                  "example": " "
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "접근 방식 오류"
          },
          "401": {
            "description": "유효하지 않은 아이디 또는 비밀번호"
          },
          "500": {
            "description": "로그인 정보 불러오기 실패"
          }
        }
      }
    },
    "/user/logout": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "로그아웃",
        "description": "현재 로그인된 사용자의 로그아웃을 수행",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "로그아웃 정보 불러오기 실패"
          }
        }
      }
    },
...
  }
}